split Logs page into separate LogsPreviewer component
